Quite a bit has changed.  I hope I can remember everything.
(1) I've removed as.data.list methods for lists and data frames and replaced them with a default method, which is almost exactly the same as the old data.list function.

(2) data.list is now largely a wrapper for as.data.list.default, but with some better naming conventions (suggested by Ben Bolker).

(3) I added a method to 'transpose' data lists, which is just a wrapper of aperm.data.list in which the perm argument is rev(1:n) where n is the number of dimensions.  I also added this same default perm argument for calls to aperm.data.list with missing perm arguments.

Hope that's everything.  Oh...I should probably say that I also started a bit of work on combining data lists created with different methods for as.data.list but didn't come up with anything work version controlling yet.  Its saved in another directory called 'oldversion'. 

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/multitable@29 edb9625f-4e0d-4859-8d74-9fd3b1da38cb